requirements.txt
discord.py
python-dotenv
transformers
torch
openai>=1.0.0
Pillow